Move CVX quadratic problems to correct folder
- Move CVXQP1, CVXQP2, CVXQP3 from _constrained_minimisation to _quadratic_problems
- These problems inherit from AbstractConstrainedQuadraticProblem
- Update all imports in __init__.py files accordingly
- All CVX problems now correctly located based on their inheritance hierarchy
